pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "puppet-ai"
version = "0.1.0"
description = "Full computer control for AI agents — see, click, type, scroll via MCP"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [{name = "Daniel Starkov"}]
dependencies = [
    "mcp>=1.0.0",
    "numpy>=1.24",
    "Pillow>=10.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "click>=8.0",
    "pyautogui>=0.9",
    "pyperclip>=1.8",
]

[project.urls]
Homepage = "https://github.com/transoff/puppet-ai"
Repository = "https://github.com/transoff/puppet-ai"
Issues = "https://github.com/transoff/puppet-ai/issues"

[project.optional-dependencies]
anthropic = ["anthropic>=0.40"]
gemini = ["google-genai>=1.0"]
openai = ["openai>=1.50"]
apple-vision = ["pyobjc-framework-Vision>=10.0", "pyobjc-framework-ScreenCaptureKit>=10.0"]
mlx = ["mlx>=0.20", "mlx-vlm>=0.1"]
all = ["puppet-ai[anthropic,gemini,openai,apple-vision,mlx]"]
dev = ["pytest>=8.0", "pytest-asyncio>=0.24", "ruff>=0.5"]

[project.scripts]
puppet-ai = "puppet_ai.cli:main"

[project.entry-points."puppet_ai.providers"]

[tool.hatch.build.targets.wheel]
packages = ["src/puppet_ai"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
